feat: 84 create cap records (#98)

+++ backend/test/conftest.py
@@ -2,7 +2,7 @@ import json
 import logging
 import os
 import sys
-from typing import Dict, Generator
+from typing import Generator
 
 import pytest
 import sqlmodel
@@ -39,21 +39,28 @@ def set_env():
 # used for postgres testing
 @pytest.fixture(scope="session")
 def db_pg_engine(set_env):
-    engine = create_engine(
-        "postgresql+psycopg2://"
-        + f"{os.environ.get('POSTGRES_USER')}:"
-        + f"{os.environ.get('POSTGRES_PASSWORD')}@"
-        + f"{os.environ.get('POSTGRES_HOST')}:"
-        f"{os.environ.get('POSTGRES_PORT')}/"
-        f"{os.environ.get('POSTGRES_DATABASE')}"
-    )
+    LOGGER.debug(f"postgres param: {os.getenv('USE_POSTGRES')}")
+    # if the env var USE_POSTGRES is not set, then use the sqlite database
+    if ( os.getenv("USE_POSTGRES_FOR_TESTS") is None) or (os.getenv("USE_POSTGRES_FOR_TESTS").lower() != "true"):
+        engine = None
+    else:
+        LOGGER.debug("creating postgres database")
+        engine = create_engine(
+            "postgresql+psycopg2://"
+            + f"{os.environ.get('POSTGRES_USER')}:"
+            + f"{os.environ.get('POSTGRES_PASSWORD')}@"
+            + f"{os.environ.get('POSTGRES_HOST')}:"
+            f"{os.environ.get('POSTGRES_PORT')}/"
+            f"{os.environ.get('POSTGRES_DATABASE')}"
+        )
 
     # session_local = sessionmaker(bind=engine)
     # test_db = session_local()
     #test_db = sqlmodel.Session(engine)
 
     yield engine
-    engine.dispose()
+    if engine is not None:
+        engine.dispose()
 
 
 @pytest.fixture(scope="session")
@@ -63,27 +70,53 @@ def db_pg_session(db_pg_connection: sqlmodel.Session):
 
 
 @pytest.fixture(scope="session")
-def db_sqllite_engine() -> Generator[Engine, None, None]:
+def db_sqllite_engine(alert_level_data) -> Generator[Engine, None, None]:
     # should re-create the database every time the tests are run, the following
     # line ensure database that maybe hanging around as a result of a failed
     # test is deleted
-    if os.path.exists("./test_db.db"):
-        LOGGER.debug("remove the database: ./test_db.db'")
-        os.remove("./test_db.db")
 
-    SQLALCHEMY_DATABASE_URL = "sqlite:///./test_db.db"
-    LOGGER.debug(f"SQL Alchemy URL: {SQLALCHEMY_DATABASE_URL}")
-    execution_options = {"schema_translate_map": {"py_api": None}}
+    # only need to do something if we are going to use the sqllite database
+    engine = None
+    if ( os.getenv("USE_POSTGRES_FOR_TESTS") is None) or (os.getenv("USE_POSTGRES_FOR_TESTS").lower() != "true"):
+        if os.path.exists("./test_db.db"):
+            LOGGER.debug("remove the database: ./test_db.db'")
+            os.remove("./test_db.db")
 
-    engine: Engine = create_engine(
-        SQLALCHEMY_DATABASE_URL,
-        connect_args={"check_same_thread": False},
-        execution_options=execution_options,
-    )
+        SQLALCHEMY_DATABASE_URL = "sqlite:///./test_db.db"
+        LOGGER.debug(f"SQL Alchemy URL: {SQLALCHEMY_DATABASE_URL}")
+        execution_options = {"schema_translate_map": {"py_api": None}}
 
-    sqlmodel.SQLModel.metadata.create_all(engine)
+        engine: Engine = create_engine(
+            SQLALCHEMY_DATABASE_URL,
+            connect_args={"check_same_thread": False},
+            execution_options=execution_options,
+        )
 
-    LOGGER.debug(f"engine type: {type(engine)}")
+        sqlmodel.SQLModel.metadata.create_all(engine)
+
+        LOGGER.debug(f"engine type: {type(engine)}")
+
+        # for sqllite database we need to load the data
+            
+        # loading basin data
+        LOGGER.info("loading basin data into test database")
+        session = sqlmodel.Session(engine)
+        basin_file = os.path.join(
+            os.path.dirname(__file__), "..", "alembic", "data", "basins.json"
+        )
+        LOGGER.debug(f"src file for basin data: {basin_file}")
+        with open(basin_file, "r") as json_file:
+            basins_data = json.load(json_file)
+
+        for basin in basins_data:
+            basin = Basins(basin_name=basin["basin_name"])  # noqa: F405
+            session.add(basin)
+        # loading alert level data
+        for alert in alert_level_data:
+            alert_level = Alert_Levels(alert_level=alert["alert_level"])  # noqa: F405
+            session.add(alert_level)
+        session.commit()
+
     yield engine
 
     # dropping all objects in the test database and...
@@ -93,7 +126,7 @@ def db_sqllite_engine() -> Generator[Engine, None, None]:
         # os.remove("./test_db.db")
 
 @pytest.fixture(scope="session")
-def alert_level_data() -> Generator[Alert_Levels_Read, None, None]:
+def alert_level_data() -> Generator[Alert_Levels_Read, None, None]:  # noqa: F405
     """
     reads the alert levels json file and loads to a List[Dict] structure.
 
@@ -114,39 +147,23 @@ def alert_level_data() -> Generator[Alert_Levels_Read, None, None]:
     yield alert_level_data
 
 @pytest.fixture(scope="session")
-def db_test_load_data(db_sqllite_engine, alert_level_data):
-    """
-    loads test data
-    """
-    engine = db_sqllite_engine
-    LOGGER.info("loading basin data into test database")
-    # loading basin data
-    session = sqlmodel.Session(engine)
-    basin_file = os.path.join(
-        os.path.dirname(__file__), "..", "alembic", "data", "basins.json"
-    )
-    LOGGER.debug(f"src file: {basin_file}")
-
-    with open(basin_file, "r") as json_file:
-        basins_data = json.load(json_file)
-
-    for basin in basins_data:
-        basin = Basins(basin_name=basin["basin_name"])  # noqa: F405
-        session.add(basin)
-    # loading alert level data
-    for alert in alert_level_data:
-        alert_level = Alert_Levels(alert_level=alert["alert_level"])  # noqa: F405
-        session.add(alert_level)
-    session.commit()
+def db_engine(db_sqllite_engine, db_pg_engine):
+    if db_pg_engine is not None:
+        LOGGER.info("using postgres database for tests")
+        engine = db_pg_engine
+    else:
+        LOGGER.info("using sqllite database for tests")
+        engine = db_sqllite_engine
+    yield engine
+    engine.dispose()
 
 # db_pg_engine db_sqllite_engine
 @pytest.fixture(scope="session")
 # def db_test_connection(db_pg_engine):
 #     engine = db_pg_engine
-def db_test_connection(db_test_load_data, db_sqllite_engine):
-    engine = db_sqllite_engine
+def db_test_connection(db_engine):
 
-    session = sqlmodel.Session(engine)
+    session = sqlmodel.Session(db_engine)
 
     yield session
     session.rollback()
@@ -174,12 +191,9 @@ def mock_access_token():
     yield token
 
 @pytest.fixture(scope="function")
-def test_client_fixture(db_test_connection, mock_access_token) -> Generator[TestClient, None, None]:
-    """returns a requests object of the current app,
-    with the objects defined in the model created in it.
+def test_app_with_auth(db_test_connection, mock_access_token):
+#def test_client_fixture(db_test_connection, mock_access_token) -> Generator[TestClient, None, None]:
 
-    :rtype: starlette.testclient
-    """
     token = mock_access_token
 
     def get_db() -> Generator[sqlmodel.Session, None, None]:
@@ -192,16 +206,35 @@ def test_client_fixture(db_test_connection, mock_access_token) -> Generator[Test
     def get_current_user():
         LOGGER.debug("current user called")
         return token
-
-    # reset to default database which points to postgres container
+       
     app.dependency_overrides[src.db.session.get_db] = get_db
     app.dependency_overrides[oidcAuthorize.authorize] = lambda: authorize()
     app.dependency_overrides[oidcAuthorize.get_current_user] = (
-        lambda: get_current_user()
+                lambda: get_current_user()
     )
 
-    yield TestClient(app)
+
+    yield app
+    app.dependency_overrides = {}
+
+
+@pytest.fixture(scope="function")
+def test_client_fixture( test_app_with_auth) -> Generator[TestClient, None, None]:
+    """returns a requests object of the current app,
+    with the objects defined in the model created in it.
+
+    :rtype: starlette.testclient
+    """
+
+    yield TestClient(test_app_with_auth)
 
     # reset other dependency override back to app default in each test
     # during test case teardown.
-    app.dependency_overrides = {}
+
+
+@pytest.fixture(scope="function")
+def test_client_fixture_w_alert_cap( test_app_with_auth) -> Generator[TestClient, None, None]:
+
+    def get_db() -> Generator[sqlmodel.Session, None, None]:
+        yield db_test_connection
+
